factor platform dependent code into separate file
